I've added a lot of code to my CSS style file.
I'm still having some difficulty translating from my CSS file to my finished product.  The link is sound.  But sometimes I make a command and nothing happens....often.  I've got a lot of code in there that I've established and it does nothing.  I've been piecing through trying to find out why.  It's tedious and I hadn't even begun to finish when my time ran out.  But I was on a bit of a roll.  I'm not frustrated but I am wondering why it's taking me so very long.